# restfull/FEMS/swagger.yaml
openapi: 3.0.3
info:
  title: TZW LTD Fire Extinguisher Management API
  version: 1.0.0
  description: |
    Consolidated API documentation for the FEMS backend services behind the API gateway.

    Covered services:
    - Authentication and user administration
    - Facility and customer management
    - Fire extinguisher registration and tracking
    - Inspection scheduling and recording
    - Maintenance logging
    - Notifications and escalations
    - Reporting and audit access
servers:
  - url: http://localhost:5000/api
    description: Local API gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit10Param:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Limit20Param:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    MaintenanceCreateRequest:
      type: object
      required:
        [extinguisherId, serviceDate, serviceCompany, technicianName, actionTaken, issuesIdentified]
      properties:
        extinguisherId:
          type: string
          format: uuid
        serviceDate:
          type: string
          format: date
        serviceCompany:
          type: string
        technicianName:
          type: string
        actionTaken:
          type: string
        issuesIdentified:
          type: string
        recommendations:
          type: string
          nullable: true
        nextServiceDate:
          type: string
          format: date
          nullable: true
        cost:
          type: number
          default: 0
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
          default: completed
    MaintenanceStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
...
    ResolveEscalationRequest:
      type: object
      properties:
        notes:
          type: string
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /maintenance:
    post:
      tags: [Maintenance]
      summary: Create a maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCreateRequest'
      responses:
        '201':
          description: Maintenance record created
    get:
      tags: [Maintenance]
      summary: List maintenance records
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - name: extinguisherId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
      responses:
        '200':
          description: Maintenance list
  /maintenance/{id}:
    get:
      tags: [Maintenance]
      summary: Get a maintenance record by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Maintenance details
    put:
      tags: [Maintenance]
      summary: Update a maintenance record by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCreateRequest'
      responses:
        '200':
          description: Maintenance record updated
  /maintenance/{id}/status:
    patch:
      tags: [Maintenance]
      summary: Update maintenance status
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceStatusRequest'
      responses:
        '200':
          description: Maintenance status updated
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit20Param'
        - name: unread
          in: query
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Notification list
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark one notification as read
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Notification marked as read
...
  /notifications/escalations:
    get:
      tags: [Escalations]
      summary: List escalation records
      description: Admin and inspector roles only
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Escalation list
  /notifications/escalations/{id}/resolve:
    patch:
      tags: [Escalations]
      summary: Resolve an escalation
      description: Admin and inspector roles only
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveEscalationRequest'
      responses:
        '200':
          description: Escalation resolved
...
  /reports/maintenance:
    get:
      tags: [Reports]
      summary: Maintenance report
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        '200':
          description: Report payload or downloadable export
...